apps/ai_engine/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Configuration
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# AI/ML Libraries
openai==1.54.0
google-cloud-vision==3.7.0
google-cloud-documentai==2.27.0
azure-ai-formrecognizer==3.3.2
pytesseract==0.3.13
Pillow==10.4.0

# HTTP Client
httpx==0.27.0
requests==2.32.3

# Utilities
python-dateutil==2.9.0
python-jose[cryptography]==3.3.0

# Logging
structlog==24.1.0

# Optional: Database support (if needed)
# sqlalchemy==2.0.35
# alembic==1.13.2

# Optional: Redis support (if needed)
# redis==5.1.0
# redis-py==5.1.0


# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
